Implement background refresher
